+++ winit/src/event_loop.rs
@@ -14,16 +14,11 @@ use std::os::unix::io::{AsFd, AsRawFd, BorrowedFd, RawFd};
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::{error, fmt};
 
-#[cfg(not(web_platform))]
-use std::time::{Duration, Instant};
-#[cfg(web_platform)]
-use web_time::{Duration, Instant};
-
 use crate::error::EventLoopError;
 use crate::{event::Event, monitor::MonitorHandle, platform_impl};
 
 #[doc(inline)]
-pub use winit_core::event_loop::AsyncRequestSerial;
+pub use winit_core::event_loop::{AsyncRequestSerial, ControlFlow};
 
 /// Provides a way to retrieve events from the system and from the windows that were registered to
 /// the events loop.
@@ -141,50 +136,6 @@ impl<T> fmt::Debug for EventLoop<T> {
 impl fmt::Debug for EventLoopWindowTarget {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.pad("EventLoopWindowTarget { .. }")
-    }
-}
-
-/// Set through [`EventLoopWindowTarget::set_control_flow()`].
-///
-/// Indicates the desired behavior of the event loop after [`Event::AboutToWait`] is emitted.
-///
-/// Defaults to [`Wait`].
-///
-/// [`Wait`]: Self::Wait
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq)]
-pub enum ControlFlow {
-    /// When the current loop iteration finishes, immediately begin a new iteration regardless of
-    /// whether or not new events are available to process.
-    Poll,
-
-    /// When the current loop iteration finishes, suspend the thread until another event arrives.
-    #[default]
-    Wait,
-
-    /// When the current loop iteration finishes, suspend the thread until either another event
-    /// arrives or the given time is reached.
-    ///
-    /// Useful for implementing efficient timers. Applications which want to render at the display's
-    /// native refresh rate should instead use [`Poll`] and the VSync functionality of a graphics API
-    /// to reduce odds of missed frames.
-    ///
-    /// [`Poll`]: Self::Poll
-    WaitUntil(Instant),
-}
-
-impl ControlFlow {
-    /// Creates a [`ControlFlow`] that waits until a timeout has expired.
-    ///
-    /// In most cases, this is set to [`WaitUntil`]. However, if the timeout overflows, it is
-    /// instead set to [`Wait`].
-    ///
-    /// [`WaitUntil`]: Self::WaitUntil
-    /// [`Wait`]: Self::Wait
-    pub fn wait_duration(timeout: Duration) -> Self {
-        match Instant::now().checked_add(timeout) {
-            Some(instant) => Self::WaitUntil(instant),
-            None => Self::Wait,
-        }
     }
 }
