--- winit/src/keyboard.rs
@@ -1,4 +1,0 @@
-//! Types related to the keyboard.
-
-#[doc(inline)]
-pub use winit_core::keyboard::*;
